Minor CR fixes - changed downlaod filename, refactored some names
