+++ monkey/monkey_island/cc/services/telemetry/processing/processing.py
