+++ monkey/monkey_island/cc/services/telemetry/zero_trust_tests/antivirus_existence.py
@@ -5,76 +5,7 @@ from common.data.zero_trust_consts import EVENT_TYPE_MONKEY_LOCAL, EVENT_TYPE_IS
 from monkey_island.cc.models import Monkey
 from monkey_island.cc.models.zero_trust.event import Event
 from monkey_island.cc.models.zero_trust.finding import Finding
-
-ANTI_VIRUS_KNOWN_PROCESS_NAMES = [
-    u"AvastSvc.exe",
-    u"AvastUI.exe",
-    u"avcenter.exe",
-    u"avconfig.exe",
-    u"avgcsrvx.exe",
-    u"avgidsagent.exe",
-    u"avgnt.exe",
-    u"avgrsx.exe",
-    u"avguard.exe",
-    u"avgui.exe",
-    u"avgwdsvc.exe",
-    u"avp.exe",
-    u"avscan.exe",
-    u"bdagent.exe",
-    u"ccuac.exe",
-    u"egui.exe",
-    u"hijackthis.exe",
-    u"instup.exe",
-    u"keyscrambler.exe",
-    u"mbam.exe",
-    u"mbamgui.exe",
-    u"mbampt.exe",
-    u"mbamscheduler.exe",
-    u"mbamservice.exe",
-    u"MpCmdRun.exe",
-    u"MSASCui.exe",
-    u"MsMpEng.exe",
-    u"rstrui.exe",
-    u"spybotsd.exe",
-    u"zlclient.exe",
-    u"SymCorpUI.exe",
-    u"ccSvcHst.exe",
-    u"ccApp.exe",
-    u"LUALL.exe",
-    u"SMC.exe",
-    u"SMCgui.exe",
-    u"Rtvscan.exe",
-    u"LuComServer.exe",
-    u"ProtectionUtilSurrogate.exe",
-    u"ClientRemote.exe",
-    u"SemSvc.exe",
-    u"SemLaunchSvc.exe",
-    u"sesmcontinst.exe",
-    u"LuCatalog.exe",
-    u"LUALL.exe",
-    u"LuCallbackProxy.exe",
-    u"LuComServer_3_3.exe",
-    u"httpd.exe",
-    u"dbisqlc.exe",
-    u"dbsrv16.exe",
-    u"semapisrv.exe",
-    u"snac64.exe",
-    u"AutoExcl.exe",
-    u"DoScan.exe",
-    u"nlnhook.exe",
-    u"SavUI.exe",
-    u"SepLiveUpdate.exe",
-    u"Smc.exe",
-    u"SmcGui.exe",
-    u"SymCorpUI.exe",
-    u"symerr.exe",
-    u"ccSvcHst.exe",
-    u"DevViewer.exe",
-    u"DWHWizrd.exe",
-    u"RtvStart.exe",
-    u"roru.exe",
-    u"WSCSAvNotifier"
-]
+from monkey_island.cc.services.telemetry.zero_trust_tests.known_anti_viruses import ANTI_VIRUS_KNOWN_PROCESS_NAMES
 
 
 def test_antivirus_existence(telemetry_json):
